Transfert toutes les methodes de test concernant le shopping cart dans ShoppingCartTest + complétion de ces test
